+++ tests/test_sim.py
@@ -10,19 +10,23 @@ import pandas as pd
 import numpy as np
 import pylab as pl
 
+
 def ret(*values):
     return values if __name__ == 'main' else None
- 
+
+
 def test_hiv_sim(n_agents=500):
     sc.heading('Test simplest possible HIV sim ')
 
     hiv = sti.HIV(
-        beta={'structuredsexual': [0.05, 0.25], 'maternal': [0.05, 0.]},
+        beta_m2f=0.05,
+        beta_m2c=0.1,
         init_prev=0.05,
     )
     pregnancy = ss.Pregnancy(fertility_rate=10)
     death = ss.Deaths(death_rate=10)
-    sexual = sti.StructuredSexual()
+    sexual = sti.StructuredSexual(recall_prior=True)
+    prior = sti.PriorPartners()
     maternal = ss.MaternalNet()
     testing = sti.HIVTest(test_prob_data=0.2, start=2000)
     art = sti.ART(coverage_data=pd.DataFrame(index=np.arange(2000, 2021), data={'p_art': np.linspace(0, 0.9, 21)}))
@@ -33,7 +37,7 @@ def test_hiv_sim(n_agents=500):
         dur=40,
         n_agents=n_agents,
         diseases=hiv,
-        networks=[sexual, maternal],
+        networks=[sexual, prior, maternal],
         demographics=[pregnancy, death],
         interventions=[testing, art, vmmc]
     )